refactor(semesters): simplify SemesterPage and enhance loading state handling
