+++ apps/backend/src/payload.config.ts
@@ -32,7 +32,9 @@ export default buildConfig({
     admin: "/payload/admin",
     api: "/payload/api",
   },
-  cors: [process.env.NEXT_PUBLIC_URL],
+  cors: {
+    origins: [process.env.NEXT_PUBLIC_URL],
+  },
   admin: {
     user: Admin.slug,
     importMap: {
