+++ apps/frontend/src/app/semesters/semester-content.tsx
@@ -1,16 +1,39 @@
 "use client"
 
 import { GetSemesterResponseSchema } from "@repo/shared"
-import { useSuspenseQuery } from "@tanstack/react-query"
-import { DataList, EmptyState, VStack } from "@yamada-ui/react"
+import { useQuery } from "@tanstack/react-query"
+import {
+  DataList,
+  DataListDescription,
+  DataListItem,
+  DataListTerm,
+  EmptyState,
+  Skeleton,
+  Text,
+  VStack,
+} from "@yamada-ui/react"
 import { useQueryState } from "nuqs"
 import { apiClient } from "@/lib/api/client"
 import { parsers } from "./search"
 
+const terms = [
+  "Name",
+  "Start Date",
+  "End Date",
+  "Break Start",
+  "Break End",
+  "Booking Open Day",
+  "Booking Open Time",
+]
+
 export default function SemesterContent() {
   const [id] = useQueryState("id", parsers.id)
 
-  const { data: semester, isError } = useSuspenseQuery({
+  const {
+    data: semester,
+    isError,
+    isLoading,
+  } = useQuery({
     queryKey: ["semester", id],
     queryFn: async () => {
       if (!id) return null
@@ -18,7 +41,7 @@ export default function SemesterContent() {
         "semester",
         id,
       ])
-      return data?.data
+      return data?.data ?? null
     },
   })
 
@@ -26,7 +49,22 @@ export default function SemesterContent() {
     <>
       {!id ? (
         <EmptyState title="Please select a semester to see details" />
-      ) : isError || !semester ? (
+      ) : isLoading ? (
+        <DataList col={2}>
+          {terms.map((term) => (
+            <DataListItem key={term}>
+              <DataListTerm>{term}</DataListTerm>
+              <DataListDescription>
+                <Skeleton>
+                  <Text>Description</Text>
+                </Skeleton>
+              </DataListDescription>
+            </DataListItem>
+          ))}
+        </DataList>
+      ) : isError ? (
+        <EmptyState title="Error loading semester" />
+      ) : !semester ? (
         <EmptyState title="Semester not found" />
       ) : (
         (() => {
